requirements.txt
# Core Data Science Libraries
numpy==1.26.4
pandas==2.2.0
scipy==1.12.0

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
tensorflow==2.15.0
torch==2.1.2

# Statistical Analysis
statsmodels==0.14.1

# NLP and Sentiment Analysis
nltk==3.8.1
transformers==4.37.2
vaderSentiment==3.3.2

# Data Acquisition
requests==2.31.0
yfinance==0.2.36
alpha-vantage==2.3.1
newsapi-python==0.2.7

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Visualization
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0

# Web Framework (for dashboard)
streamlit==1.31.0
flask==3.0.1

# Utilities
python-dotenv==1.0.1
tqdm==4.66.1
joblib==1.3.2

# Configuration
pyyaml==6.0.1

# Testing
pytest==8.0.0
pytest-cov==4.1.0
